% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweetAleRt.R
\name{confirmDialog}
\alias{confirmDialog}
\title{Show a confirmation dialog based on sweetalert on the clients screen.}
\usage{
confirmDialog(inputId, questionTitle = NULL, questionText = NULL,
  questionIcon = NULL, dangerMode = F, acceptTitle = NULL,
  acceptText = NULL, acceptIcon = NULL, acceptTimer = NULL,
  declineTitle = NULL, declineText = NULL, declineIcon = NULL,
  declineTimer = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{ID of the input under which the acceptance will be available}

\item{questionTitle}{Title of the confirmation dialog}

\item{questionText}{Text to display in the confirmation dialog}

\item{questionIcon}{Icon of the confirmation dialog. Available are
warning, error, success, info}

\item{dangerMode}{Enable danger mode. The danger mode sets the focus
to the cancel button instead of accept.}

\item{acceptTitle}{Title of the accept alert}

\item{acceptText}{Text to display in the accept alert}

\item{acceptIcon}{Icon of the accept alert. Available are
warning, error, success, info}

\item{acceptTimer}{Timout for the accept alert to disappear}

\item{declineTitle}{Title of the decline alert}

\item{declineText}{Text to display in the decline alert}

\item{declineIcon}{Icon of the decline alert. Available are
warning, error, success, info}

\item{declineTimer}{Timout for the decline alert to disappear}

\item{session}{Shiny session object}
}
\value{
HTML-Tags
}
\description{
This function creates and displays a confirmation dialog based on
sweetalert on the clients screen.
All arguments are inspired by \link{https://sweetalert.js.org/docs/}
}
\examples{
shinyApp(
ui = SweetAleRt::includeSweetAlert(),
server = function(session, input, output){
   SweetAleRt::confirmDialog("confirm", "Question", "Do you want to continue?", "warning",
   acceptText = "continuing...", acceptIcon = "success", acceptTimer = 1000)
})
}
\seealso{
\link{https://sweetalert.js.org/docs/}
}
