% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweetAleRt.R
\name{sweetAlert}
\alias{sweetAlert}
\title{Show a sweetalert on the clients screen.}
\usage{
sweetAlert(text, title = NULL, icon = NULL, button = NULL,
  buttons = NULL, content = NULL, closeOnClickOutside = T,
  closeOnEsc = T, dangerMode = F, timer = NULL,
  session = getDefaultReactiveDomain())
}
\arguments{
\item{text}{Text to display in the alert}

\item{title}{Title of the alert}

\item{icon}{Icon of the alert. Available are warning, error,
success, info}

\item{button}{Standart button of the alert}

\item{buttons}{Custom buttons of the alert}

\item{content}{Custom inputs in the alert}

\item{closeOnClickOutside}{Boolean, close the alert on a click outside
of the alert}

\item{closeOnEsc}{Boolean, close the alert by pressing the escape key}

\item{dangerMode}{Enable danger mode. The danger mode sets the focus
to the cancel button instead of accept.}

\item{timer}{Timeout to automatically close the alert}

\item{session}{Shiny session object}
}
\value{
HTML-Tags
}
\description{
This function creates and displays a sweetalert on the clients screen.
All arguments are inspired by \link{https://sweetalert.js.org/docs/}
}
\examples{
shinyApp(
ui = SweetAleRt::includeSweetAlert(),
server = function(session, input, output){
   SweetAleRt::sweetAlert("Hello World!")
})
}
\seealso{
\link{https://sweetalert.js.org/docs/}
}
